Tribal_Knowledge/pyproject.toml
[tool.poetry]
name = "tribal-knowledge"
version = "0.1.0"
description = "streamlit app for assessing IT risk example"
authors = ["michael pearmain"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
numpy = "^1.26.4"
pandas = "^2.2.0"
scipy = "^1.12.0"
plotly = "^5.19.0"
pyarrow = "^15.0.0"
arviz = "^0.17.0"
jupyterlab = "^4.1.2"
jupyter = "^1.0.0"
pymc = "^5.10.4"
graphviz = "^0.20.1"
ridgeplot = "^0.1.24"
matplotlib = "^3.8.3"
pytensor = "^2.18.6"
numba = "^0.59.0"
nutpie = "^0.9.2"
blackjax = "^1.1.1"
pygithub = "^2.2.0"
streamlit = "^1.32.2"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
